% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQUIC_R.R
\name{SQUIC}
\alias{SQUIC}
\title{Sparse Quadratic Inverse Covariance Estimation}
\usage{
SQUIC(
  Y = NULL,
  lambda = 0,
  X_pattern = NULL,
  max_iter_in = 1,
  drop_tol_in = 1e-06,
  term_tol_in = 1e-06,
  X_init = NULL,
  W_init = NULL,
  verbose_in = 0
)
}
\arguments{
\item{Y}{Input data, expects a matrix of dimension p x n, where each sample corresponds to one row of length p and a total number of n samples.}

\item{lambda_1}{Regularization parameter. It can either be a non-negative scalar. Then each entry of X will be penalized in the same way with this value. Or it can be a matrix of the same dimension as X, that is p by p, with non-negative values. Then each entry X_ij will be penalised by the value of ij-th entry of the matrix.}

\item{lambda_2}{... Optional parameter that determines if}

\item{X0}{Optimal parameter. Matrix with a sparse initial guess of the inverse covariance matrix. Defaults to the identity matrix. Needs to be given in coo format.}

\item{W0}{Optimal parameter. Matrix with a sparse initial guess of the covariance matrix. Defaults to the identity matrix. Needs to be given in coo format.}

\item{max_iter}{Maximum number of iterations.}

\item{drop_tol}{Convergence threshold. Defaults to 1e-5.}

\item{term_tol}{Convergence threshold. Defaults to 1e-5}

\item{verbose}{Integer value. Defining the verbosity of the print statistics. The higher the value the more detailed the output gets. The value range is from: ... Defaults to zero.}
}
\value{
X Estimated sparse inverse covariance matrix

W Estimated sparse inverse of the inverse covariance matrix
}
\description{
This function computes large Sparse Inverse Covariance Matrices quickly. 
It returns an estimate of the sparse covariance matrix W and its inverse X of a data matrix Y.
}
\details{
R version of SQUIC.
The higher the values are for lambda_1 the more sparsity is enforced. Possible source of how to choose lambda. 


The optional parameters X0 and W0 are for sparse initial guesses of X and W. 
They need to be given in COO-Format, therefore they each contain 3 lists. As these matrices are symmetric the function only requires the lower triangular part. If not specified they will default to identity matrices.
}
\references{
Large-scale Sparse Inverse Covariance Matrix Estimation, M.Bollhöfer, A. Eftekhari, S. Scheidegger, O. Schenk \url{https://epubs.siam.org/doi/abs/10.1137/17M1147615?journalCode=sjoce3}
}
\seealso{
\link{BigQuic}, \link{QUIC}
}
\author{

}
\keyword{Covariance}
\keyword{Estimation}
\keyword{Matrix}
\keyword{Sparse}
